+++ Utilities/Users/UserDetails.cs
@@ -7,7 +7,7 @@
         public string isBanned { get; set; }
         public string externalAppDisplayName { get; set; }
         public bool hasVerifiedBadge { get; set; }
-        public int id { get; set; }
+        public ulong id { get; set; }
         public string name { get; set; }
         public string displayName { get; set; }
     }
