+++ Utilities/Users/UserIdPost.cs
@@ -2,7 +2,7 @@
 {
     public class UserIdPost
     {
-        public int[] userIds { get; set; }
+        public ulong[] userIds { get; set; }
         public bool excludeBannedUsers { get; set; }
     }
 }
